Update models to TypeScript files
