+++ src/models/common/image.ts
@@ -1,5 +1,3 @@
-/* @flow */
-
 class Image {
     height: number;
