+++ src/models/other/track-link.ts
@@ -1,5 +1,3 @@
-/* @flow */
-
 import ExternalUrl from '../common/external-url';
 
 class TrackLink {
