+++ src/models/paging/cursor-based-paging.ts
@@ -1,5 +1,3 @@
-/* @flow */
-
 import Cursor from './cursor';
 
 class CursorPaging<T> {
