+++ src/models/paging/cursor.ts
@@ -1,5 +1,3 @@
-/* @flow */
-
 class Cursor {
     after: string;
