+++ src/models/paging/paging.ts
@@ -1,5 +1,3 @@
-/* @flow */
-
 class Paging<T> {
     href: string;
