+++ src/models/player/context.ts
@@ -1,5 +1,3 @@
-/* @flow */
-
 import ExternalUrl from '../common/external-url';
 
 class Context {
