+++ src/models/track/audio-features.ts
@@ -1,5 +1,3 @@
-/* @flow */
-
 class AudioFeatures {
     acousticness: number;
