+++ src/models/user/user-private.ts
@@ -1,15 +1,19 @@
-/* @flow */
-
 import ExternalUrl from '../common/external-url';
 import Followers from '../common/followers';
 import Image from '../common/image';
 
-class Artist {
+class PublicUser {
+    birthdate: string;
+
+    country: string;
+
+    displayName: string;
+
+    email: string;
+
     externalUrls: ExternalUrl;
 
     followers: Followers;
-
-    genres: Array<string>;
 
     href: string;
 
@@ -17,29 +21,29 @@ class Artist {
 
     images: Array<Image>;
 
-    name: string;
+    product: string;
 
-    popularity: number;
-
-    type: 'artist';
+    type: 'user';
 
     uri: string;
 
     constructor(json: any) {
         if (json) {
+            this.birthdate = json.birthdate;
+            this.country = json.country;
+            this.displayName = json.display_name;
+            this.email = json.email;
             this.externalUrls = new ExternalUrl(json.external_urls);
             this.followers = new Followers(json.followers);
-            this.genres = json.genres;
             this.href = json.href;
             this.id = json.id;
             this.images = json.images
                 .map(imageJson => new Image(imageJson));
-            this.name = json.name;
-            this.popularity = json.popularity;
+            this.product = json.product;
             this.type = json.type;
             this.uri = json.uri;
         }
     }
 }
 
-export default Artist;
+export default PublicUser;
